--- .gitattributes
@@ -1,5 +1,0 @@
-# Set default behavior to automatically normalize line endings.
-* text=auto
-
-# For LFS
-*.h5 filter=lfs diff=lfs merge=lfs -text
