--- .gitignore
@@ -1,3 +1,0 @@
-.vscode
-venv_*
-__pycache__
