+++ backend/.gitattributes
@@ -1,0 +1,5 @@
+# Set default behavior to automatically normalize line endings.
+* text=auto
+
+# For LFS
+*.h5 filter=lfs diff=lfs merge=lfs -text
