+++ backend/.gitignore
@@ -1,0 +1,3 @@
+.vscode
+venv_*
+__pycache__
