+++ backend/code/__init__.py
@@ -1,0 +1,1 @@
+from code.predict import *
