+++ backend/code/constants.py
@@ -1,0 +1,3 @@
+IMAGE_SIZE = (256, 256)
+CLASSIFIER_WEIGHTS_PATH = "static/weights/classifier_v2.h5"
+AUTOENCODER_WEIGHTS_PATH = "static/weights/autoencoder_v1.h5"
