+++ backend/static/weights/autoencoder_v1.h5
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:ff95d22080b577dec7ab3845634a8b9db2a28d7b9ca003874113c2493aee915f
+size 9534248
