+++ backend/static/weights/classifier_v2.h5
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:7352f1adddaffdba3c3333dd943ccec92f7521b56af8abb939a3453c1aefb530
+size 103970792
