--- code/model/customF1.py
@@ -1,21 +1,0 @@
-from keras import backend as K
-
-def custom_f1(y_true, y_pred):    
-    def recall_m(y_true, y_pred):
-        TP = K.sum(K.round(K.clip(y_true * y_pred, 0, 1)))
-        Positives = K.sum(K.round(K.clip(y_true, 0, 1)))
-        
-        recall = TP / (Positives+K.epsilon())    
-        return recall 
-    
-    
-    def precision_m(y_true, y_pred):
-        TP = K.sum(K.round(K.clip(y_true * y_pred, 0, 1)))
-        Pred_Positives = K.sum(K.round(K.clip(y_pred, 0, 1)))
-    
-        precision = TP / (Pred_Positives+K.epsilon())
-        return precision 
-    
-    precision, recall = precision_m(y_true, y_pred), recall_m(y_true, y_pred)
-    
-    return 2*((precision*recall)/(precision+recall+K.epsilon()))
