--- requirements.txt
@@ -1,5 +1,0 @@
-fastapi==0.82.0
-Pillow~=9.2.0
-numpy~=1.23.2
-uvicorn~=0.18.3
-keras~=2.10.0
